Fix valkey serialization and parsing

+++ tests/test_pipes_dtypes.py
@@ -977,6 +977,8 @@ def test_enforce_false(flavor: str):
     """
     Test `enforce=False` behavior.
     """ 
+    if flavor not in ('api', 'mssql', 'timescaledb', 'sqlite'):
+        return
     conn = conns[flavor]
     pipe = mrsm.Pipe('test', 'enforce', instance=conn)
     pipe.delete()
@@ -984,6 +986,7 @@ def test_enforce_false(flavor: str):
         'test', 'enforce',
         instance=conn,
         enforce=False,
+        dtypes={'dt': 'datetime64[ns]'},
         columns={
             'datetime': 'dt',
         },
